Allow projects to opt in to running Minitest cops
As a former `RSpec` user it would be nice to be able to opt-in to some
community syntax guidance when it comes to `Minitest`.

Since the `.rubocop.yml` is configured with the
`AllCops.DisabledByDefault` set to `true` this commit doesn’t enforce
any new cops. Instead projects (or individual programmers) can opt in to
having their editor lint `Minitest` without having to manipulate the
project’s `Gemfile`.